requirements.txt
absl-py==0.11.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
idna==2.10
joblib==1.0.0
kiwisolver==1.3.1
llvmlite==0.35.0
Markdown==3.3.3
matplotlib==3.3.3
numba==0.52.0
numpy==1.19.4
oauthlib==3.1.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.55.0
typing-extensions==3.7.4.3
umap-learn==0.4.6
urllib3==1.26.2
Werkzeug==1.0.1
